+++ src/bible/routes/chapters.nim
@@ -2,6 +2,7 @@ from std/strformat import fmt
 from std/strutils import contains
 
 import pkg/prologue
+from pkg/bibleTools import identifyBibleBook, enAbbr
 
 import bible/db/models/[
   book,
@@ -19,7 +20,7 @@ proc r_chapters*(ctx: Context) {.async.} =
     node.ifContains(all = ["doc", "book"]):
       let
         doc = node{"doc"}.getStr
-        book = node{"book"}.getStr
+        book = node{"book"}.getStr.identifyBibleBook.book.enAbbr
 
       ctx.withDoc doc:
         let chapters = doc.getChaptersQnt(book)