+++ src/bible/routes/verses.nim
@@ -3,6 +3,7 @@ from std/strutils import contains
 import pkg/prologue
 
 from pkg/util/forStr import tryParseInt
+from pkg/bibleTools import identifyBibleBook, enAbbr
 
 import bible/db/models/[
   verse,
@@ -21,7 +22,7 @@ proc r_verses*(ctx: Context) {.async.} =
     node.ifContains(all = ["doc", "book", "chapter"]):
       let
         doc = node{"doc"}.getStr
-        book = node{"book"}.getStr
+        book = node{"book"}.getStr.identifyBibleBook.book.enAbbr
         chapter = node{"chapter"}.getStr.tryParseInt 0
 
       ctx.withDoc doc: